requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
numpy-stl==3.1.1
numpy==1.24.4
jinja2==3.1.3
python-multipart==0.0.9
itsdangerous==2.1.2
